requirements.txt
# Self Analysis AI - Dependencies

streamlit>=1.28.0
pydantic>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
altair>=5.0.0
psycopg2-binary>=2.9.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
